split terraform client to its own file
